Add files via upload
On lines 6 and 7, in order to pull the user's investment amount and the current price of BTC, I made them an input function. 
Once the user inputs the numerical values, all the user sees on their end is the total price of their investment in USD in real-time (current_time).